+++ src/scenarioInfo.h
@@ -36,6 +36,7 @@ public:
     static std::vector<string> getCategories();
     static const std::vector<ScenarioInfo>& getScenarios();
     static std::vector<ScenarioInfo> getScenarios(const string& category);
+    static void clearCache();
 private:
     void addKeyValue(string key, string value);
     bool addSettingOption(string key, string option, string description);